Add files via upload
This game prompts the user to enter their name, then generates a random number and asks the user to guess it. The user has 10 attempts to guess the correct number. If they succeed, the game prints a congratulations message. If the user runs out of guesses, the game prints the user ran out of guesses and did not win.